Introduce <script type=webbundle>
Introduce <script>-based API for subresource loading with Web Bundles.

See the design doc [1] for the motivation of switching from
<link>-based API to <script>-based API. The explainer [2] was already
updated to use <script>-based API.

This feature is guarded by `SubresourceWebBundles` flag.

We eventually drop the <link rel=webbundle> support and remove the
<link>-based API code once we can confirm <script>-based API can be
used as a replacement of <link>-based API.

This CL should be considered as the first step to switch to
<script>-based API. There are still gaps between <link>-based API and
<script>-based API, which will be addressed later [3].

This CL intentionally adds a very minimum test for <script
type=webbundle>. Tests will be added in follow-up CLs. The plan is:

1. We will convert the existing tests incrementally, replacing
   <link>-based API with new <script>-based API. For example, for WPT,
   we will convert:

   from: wpt/web-bundle/subresource-loading/link-*.html
   to: wpt/web-bundle/subresource-loading/script-.html

2. If we find an issue of the implementation or find a missing feature
   by rewriting a test, we'll fix or update our implementation.

3. Continue until we can convert all existing tests.

4. Once we finish converting all tests, we can switch to
   <script>-based API and be ready to drop <link>-based API.

We'll also add tests which are specific to <script>-based API as
necessary. These efforts should be tracked by crbug.com/1245166.

[1]: https://docs.google.com/document/d/1q_SodTcLuwya4cXt1gIRaVrkiaBfwWyPvkY1fqRKkgM/edit?usp=sharing&resourcekey=0-dqrFOGVCYsg8WRZ4RFgwuw
[2]: https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md
[3]: https://github.com/WICG/webpackage/issues/670

Bug: 1245166
Change-Id: I5109b6e692baf10fd1d8a31a31d93176d4dc4ad2